+++ db/migrate/20240302164158_create_product_variants.rb
@@ -1,0 +1,11 @@
+class CreateProductVariants < ActiveRecord::Migration[7.1]
+  def change
+    create_table :product_variants do |t|
+      t.string :name
+      t.text :description
+      t.integer :year
+
+      t.timestamps
+    end
+  end
+end