+++ db/migrate/20240302164235_add_product_id_to_product_variant.rb
@@ -1,0 +1,5 @@
+class AddProductIdToProductVariant < ActiveRecord::Migration[7.1]
+  def change
+    add_column :product_variants, :product_id, :bigint
+  end
+end