+++ db/migrate/20240303011010_adds_price_to_product_variant.rb
@@ -1,0 +1,6 @@
+class AddsPriceToProductVariant < ActiveRecord::Migration[7.1]
+  def change
+    add_column :product_variants, :price, :decimal, precision: 12, scale: 4
+    add_column :product_variants, :price_currency, :string
+  end
+end